requirements.txt
Pillow
screeninfo
